requirements.txt
Pillow
screeninfo
